pyproject.toml
[project]
name = "helphub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.124.4",
    "fastmcp>=2.14.0",
    "jinja2>=3.1.6",
    "langchain>=1.1.3",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.3",
    "langgraph>=1.0.5",
    "passlib[argon2]>=1.7.4",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.1",
]
